处理算分问题

+++ src/main/java/com/trs/gov/kpi/entity/responsedata/WkOneSiteScoreResponse.java
@@ -14,13 +14,13 @@ public class WkOneSiteScoreResponse {
 
     private double total;//综合评分
 
-    private Integer invalidLink;//链接可用性分数
+    private double invalidLink;//链接可用性分数
 
-    private Integer contentError;//内容检测分数
+    private double contentError;//内容检测分数
 
-    private Integer overSpeed;//访问速度分数
+    private double overSpeed;//访问速度分数
 
-    private Integer updateContent;//网站更新分数
+    private double updateContent;//网站更新分数
 
     private Date checkTime;//入库时间
 }
